test(core): split subagent lifecycle helpers
